合并MVStore bugfixes 1.concurrent updates could result in a "Too many open transactions" exception. 2.the maximum cache size was artificially limited to 2 GB (due to an integer overflow).

+++ lealone-storage/mvstore/src/main/java/org/lealone/mvstore/MVStore.java
@@ -311,7 +311,7 @@ public class MVStore {
         o = config.get("cacheSize");
         int mb = o == null ? 16 : (Integer) o;
         if (mb > 0) {
-            int maxMemoryBytes = mb * 1024 * 1024;
+            long maxMemoryBytes = mb * 1024L * 1024L;
             int segmentCount = 16;
             int stackMoveDistance = 8;
             cache = new CacheLongKeyLIRS<Page>(maxMemoryBytes, segmentCount, stackMoveDistance);
